+++ requirements.txt
@@ -3,3 +3,4 @@ psycopg2-binary
 dynaconf
 google-cloud-storage
 git-python
+docopt
